+++ Tests/Shared/FixtureSourceGraphTestCase.swift
@@ -17,7 +17,7 @@ class FixtureSourceGraphTestCase: SourceGraphTestCase {
         configuration.retainObjcAccessible = retainObjcAccessible
         configuration.retainObjcAnnotated = retainObjcAnnotated
         configuration.indexExclude = Self.sourceFiles.subtracting([testFixturePath]).map { $0.string }
-        configuration.resetIndexExcludeMatchers()
+        configuration.resetMatchers()
 
         if !testFixturePath.exists {
             fatalError("\(testFixturePath.string) does not exist")